+++ main.rb
@@ -1,0 +1,7 @@
+require_relative 'player'
+require_relative 'question'
+require_relative 'game'
+
+
+game = Game.new()
+game.start
